First draft of message version 1
